TG-86 #done All that is missing is adapting it to the new Navigator

+++ app/src/main/java/com/potus/potus_front/API/requests/InformLocationRequest.kt
@@ -4,7 +4,7 @@ import org.jetbrains.annotations.NotNull
 
 data class InformLocationRequest(
     @NotNull()
-    var latitude: Int,
+    var latitude: Double,
     @NotNull()
-    var length: Int
+    var length: Double
 )
